pyproject.toml
[tool.poetry]
name = "bradbot"
version = "0.1.0"
description = ""
authors = ["bradtimmis <bradleytimmis@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
nextcord = "^2.0.0-alpha.6"
loguru = "^0.5.3"
python-dotenv = "^0.19.2"
aiohttp = "^3.8.1"
aiodns = "^3.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
